Move module context to internal header
