+++ src/compositor/internal_context.h
@@ -1,0 +1,50 @@
+/*
+ * waysome - wayland based window manager
+ *
+ * Copyright in alphabetical order:
+ *
+ * Copyright (C) 2014-2015 Julian Ganz
+ * Copyright (C) 2014-2015 Manuel Messner
+ * Copyright (C) 2014-2015 Marcel Müller
+ * Copyright (C) 2014-2015 Matthias Beyer
+ * Copyright (C) 2014-2015 Nadja Sommerfeld
+ *
+ * This file is part of waysome.
+ *
+ * waysome is free software: you can redistribute it and/or modify it under the
+ * terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation, either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * waysome is distributed in the hope that it will be useful, but WITHOUT ANY
+ * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+ * A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with waysome. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef __WS_COMPOSITOR_INTERNAL_CONTEXT_H__
+#define __WS_COMPOSITOR_INTERNAL_CONTEXT_H__
+
+/**
+ * A framebuffer device
+ */
+struct ws_framebuffer_device {
+    int fd; //<! The FileDescriptor
+    const char* path; //<! The path to the given filedescriptor
+};
+
+/**
+ * Internal compositor context
+ *
+ * This context holds the internal state of the compositor.
+ */
+extern struct ws_compositor_context {
+    struct ws_framebuffer_device fb;
+    struct ws_monitor* conns; //<! A linked list of ws_monitors
+} ws_comp_ctx;
+
+
+#endif // __WS_COMPOSITOR_INTERNAL_CONTEXT_H__
